requirements.txt
torch
tqdm
torchvision
numpy
opencv-python
tk
pillow
detect
matplotlib